+++ src/FilamentIO/write_data.jl
@@ -1,7 +1,9 @@
 # Write data on filament nodes -- case of scalar data (e.g. local curvature magnitude)
 function write_data_on_filaments(
-        ::Type{T}, gdata, vs::AbstractVector{<:AbstractVector}, name, refinement,
+        ::Type{T}, writer::VTKHDFFile, gdata, vs::AbstractVector{<:AbstractVector},
+        name,
     ) where {T <: Number}
+    (; refinement, fs,) = writer
     @assert T === eltype(eltype(vs))
     num_points = sum(v -> refinement * length(v) + 1, vs)  # the +1 is to include the endpoint
     dtype = HDF5.datatype(T)
@@ -13,7 +15,7 @@ function write_data_on_filaments(
     n = if refinement == 1
         write_data_unrefined(info, vs; ldims)
     else
-        write_data_refined(info, vs, refinement; ldims)
+        write_data_refined(info, vs, fs, refinement; ldims)
     end
     @assert n == num_points
 
@@ -24,8 +26,10 @@ end
 
 # Write data on filament nodes -- case of vector data (e.g. velocity)
 function write_data_on_filaments(
-        ::Type{V}, gdata, vs::AbstractVector{<:AbstractVector}, name, refinement,
+        ::Type{V}, writer::VTKHDFFile, gdata, vs::AbstractVector{<:AbstractVector},
+        name,
     ) where {V <: AbstractVector}
+    (; refinement, fs,) = writer
     @assert V === eltype(eltype(vs))
     T = eltype(V)
     @assert T <: Number
@@ -40,7 +44,7 @@ function write_data_on_filaments(
     n = if refinement == 1
         write_data_unrefined(info, vs; ldims)
     else
-        write_data_refined(info, vs, refinement; ldims)
+        write_data_refined(info, vs, fs, refinement; ldims)
     end
 
     @assert n == num_points
@@ -71,17 +75,18 @@ function write_data_unrefined(info, vs; ldims::Dims = ())
     n
 end
 
-function write_data_refined(info, vs, refinement; ldims::Dims = ())
+function write_data_refined(info, vs, fs, refinement; ldims::Dims = ())
     V = eltype(eltype(vs))  # usually SVector{3, T} or just T (scalar data)
     T = eltype(V)
     @assert T <: Number
+    @assert length(vs) == length(fs)  # 1 data vector per filament
     n = 0
     buf = Bumper.default_buffer()
-    for v ∈ vs
+    for (v, f) ∈ zip(vs, fs)
         @no_escape buf begin
             Np = refinement * length(v) + 1  # number of output points (the +1 is to close the loop)
-            Xs = @alloc(T, ldims..., Np)
-            refine_data_on_filament!(Xs, v, refinement)
+            Xs = @alloc(V, Np)
+            refine_data_on_filament!(Xs, v, f, refinement, buf)
             write_points_to_hdf5(info, Xs, n, ldims, Np)
             n += Np
         end
@@ -89,27 +94,35 @@ function write_data_refined(info, vs, refinement; ldims::Dims = ())
     n
 end
 
-function refine_data_on_filament!(Xs::AbstractArray, v::AbstractVector, refinement::Int)
-    n = firstindex(Xs)::Int - 1
-    for j ∈ eachindex(v), m ∈ 1:refinement
-        u = _interpolate_on_segment(v, j, m, refinement)
-        for w ∈ u
-            Xs[n += 1] = w
+function refine_data_on_filament!(
+        Xs::AbstractVector{T}, vs::AbstractVector{T}, f::ClosedFilament,
+        refinement::Int, buf,
+    ) where {T}
+    method = Filaments.discretisation_method(f)
+    ts = Filaments.knots(f)
+    M = Filaments.npad(method)
+    Np = length(vs)
+    @no_escape buf begin
+        data = @alloc(T, Np + 2M)
+        cs = PaddedVector{M}(data)
+        nderiv = Filaments.required_derivatives(method)
+        cderiv = ntuple(Val(nderiv)) do _
+            local data = @alloc(T, Np + 2M)
+            PaddedVector{M}(data)
         end
+        coefs = Filaments.init_coefficients(method, cs, cderiv)
+        Filaments.compute_coefficients!(coefs, vs, ts)
+        n = firstindex(Xs)::Int - 1
+        for j ∈ eachindex(vs)
+            Xs[n += 1] = vs[j]  # copy value on node
+            for m ∈ 2:refinement
+                # Copy value interpolated in-between nodes
+                ζ = (m - 1) / refinement  # in ]0, 1[
+                Xs[n += 1] = Filaments.evaluate(coefs, ts, j, ζ)
+            end
+        end
+        Xs[n += 1] = vs[begin]  # close the loop
+        @assert n == lastindex(Xs)
     end
-    for w ∈ v[end]
-        Xs[n += 1] = w  # close the loop
-    end
-    @assert n == lastindex(Xs)
     Xs
 end
-
-# Interpolate value on segment. For now we simply use linear interpolation (it's just for
-# visualisation anyways).
-function _interpolate_on_segment(v, i, m, M)
-    α = (m - 1) / M
-    (1 - α) * v[i] + α * _vnext(v, i)
-end
-
-_vnext(v::PaddedVector, i) = v[i + 1]
-_vnext(v::AbstractVector, i) = v[i == lastindex(v) ? firstindex(v) : (i + 1)]