+++ src/FilamentIO/write_points.jl
@@ -13,16 +13,15 @@ end
 
 function write_points_refined(points, fs, refinement)
     V = eltype(eltype(fs))  # usually V == SVector{3, T}
-    T = eltype(V)
     N = length(V)  # usually == 3 (works when V <: StaticArray)
     n = 0
     buf = Bumper.default_buffer()
     for f ∈ fs
         @no_escape buf begin
             Np = refinement * length(f) + 1  # number of output points (the +1 is to close the loop)
-            Xs = @alloc(T, N, Np)
+            Xs = @alloc(V, Np)
             refine_filament_coordinates!(Xs, f, refinement)
-            @assert @view(Xs[:, end]) ≈ f[end + 1]  # already includes the endpoint
+            @assert Xs[end] == f[end + 1]  # already includes the endpoint
             write_points_to_hdf5(points, Xs, n, (N,), Np)  # write data to existing HDF5 dataset
             n += Np
         end
@@ -41,26 +40,17 @@ function write_points_to_hdf5(points, Xs, n, ldims, Np)
     nothing
 end
 
-function refine_filament_coordinates!(Xs::AbstractMatrix, f::ClosedFilament, refinement::Int)
-    Xs_nodes = nodes(f)
+function refine_filament_coordinates!(Xs::AbstractVector, f::ClosedFilament, refinement::Int)
     refinement ≥ 1 || error("refinement must be ≥ 1")
-    refinement == 1 && return Xs_nodes[begin:end + 1]
-    V = eltype(f)  # usually V == SVector{3, T}
     n = firstindex(Xs, 2) - 1
-    subinds = range(0, 1; length = refinement + 1)[1:refinement]
-    for j ∈ eachindex(f), ζ ∈ subinds
-        n += 1
-        x⃗ = f(j, ζ) :: V
-        for i ∈ eachindex(x⃗)
-            Xs[i, n] = x⃗[i]
+    for j ∈ eachindex(f)
+        Xs[n += 1] = f[j]
+        for m ∈ 2:refinement
+            ζ = (m - 1) / refinement  # in ]0, 1[
+            Xs[n += 1] = f(j, ζ)
         end
     end
     @assert n == refinement * length(f)
-    let x⃗ = f(lastindex(f), 1.0)  # close the loop
-        n += 1
-        for i ∈ eachindex(x⃗)
-            Xs[i, n] = x⃗[i]
-        end
-    end
+    Xs[n += 1] = f[end + 1]  # close the loop
     Xs
 end